requirements.txt
playwright
python-dotenv
requests
pytz
aiohttp
